% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dognames.R
\name{find_dogname}
\alias{find_dogname}
\title{Find a dog name based on selected attributes}
\usage{
find_dogname(gender = NULL, breed = NULL, size = NULL, character = NULL)
}
\arguments{
\item{gender}{Character string specifying the dog's gender. NULL by default, picks a random gender name.}

\item{breed}{Character string specifying the dog's breed.  NULL by default, picks a random breed. name.}

\item{size}{Character string specifying the dog's size category.}

\item{character}{Character string specifying the dog's character trait.}
}
\value{
A single character string containing a dog name.
}
\description{
Randomly samples a dog name from the \code{dogs} dataset, optionally
conditioning on gender, breed, size, and character. If no exact match
exists, constraints are relaxed stepwise to ensure a valid result.
}
\examples{
# find a female and Pug name
find_dogname(gender = "female", breed = "Pug")
find_dogname(size = "chonker", character = "aggressive")
find_dogname()
}
